sidebar and navbar(for main page) functions created, integrated only to contest flow
