+++ gamify/src/contest/Contest.js
@@ -1,11 +1,11 @@
 import '../App.css';
 import '../index.css';
 import './Contest.css';
-import {Navbar} from "react-bootstrap";
-import {ProSidebar, Menu, MenuItem, SubMenu} from 'react-pro-sidebar';
 import 'react-pro-sidebar/dist/css/styles.css';
 import {useNavigate} from "react-router";
 import React, {useState} from "react";
+import SidebarCustom from '../sidebarAndNavbar/SidebarCustom';
+import NavbarCustom from "../sidebarAndNavbar/NavbarCustom";
 
 
 export default function Contest() {
@@ -29,10 +29,6 @@ export default function Contest() {
 
     const goBack = () => {
         navigate('/mainPage');
-    }
-
-    const logOut = () => {
-        navigate('/');
     }
 
     const handleSelection = (event) => {
@@ -72,33 +68,8 @@ export default function Contest() {
 
     return(
         <div>
-            <Navbar sticky="top" className="navbar" style={{fontFamily: "poppins"}}>
-                <div>
-                    <p className="navBarTitle" style={{display: "inline", color: "white"}}>EnergX</p>
-                    <button className="navbarButton" onClick={logOut}> Log out</button>
-                </div>
-            </Navbar>
-            <ProSidebar style={{
-                marginTop: 48,
-                position: "fixed",
-                fontFamily: "poppins",
-                sticky: "top",
-                width: "150px",
-                minWidth: "150px",
-                fontSize: "20px",
-            }} className="mySideBar">
-                <Menu iconShape="square">
-                    <MenuItem onClick={() => {}}>Menu</MenuItem>
-                    <SubMenu title="Games">
-                        <MenuItem onClick={() => navigate("/admin/users")}> Classic quiz</MenuItem>
-                        <MenuItem onClick={() => {}}> Drag & Drop</MenuItem>
-                    </SubMenu>
-                    <SubMenu title="Courses">
-                        <MenuItem onClick={() => navigate("/admin/devices")}> English</MenuItem>
-                        <MenuItem onClick={() => {}}> IT</MenuItem>
-                    </SubMenu>
-                </Menu>
-            </ProSidebar>
+            <NavbarCustom/>
+            <SidebarCustom/>
             <div className="App" style={{fontFamily:"poppins"}}>
                 <header className="myHeader">
                     <h1> Choose a subject </h1>
