+++ gamify/src/contest/ContestGettingReady.js
@@ -7,6 +7,8 @@ import 'react-pro-sidebar/dist/css/styles.css';
 import {useNavigate} from "react-router";
 import React, {useEffect} from "react";
 import {useState} from "react";
+import NavbarCustom from "../sidebarAndNavbar/NavbarCustom";
+import SidebarCustom from "../sidebarAndNavbar/SidebarCustom";
 
 export default function ContestGettingReady() {
 
@@ -24,41 +26,12 @@ export default function ContestGettingReady() {
         return () => clearInterval(interval);
     }, [timeUntilStartContest]);
 
-    const logOut = () => {
-        navigate('/');
-    }
-
     console.log("***" + JSON.parse(localStorage.getItem("numberOfPlayers")))
 
     return(
         <div>
-            <Navbar sticky="top" className="navbar" style={{fontFamily: "poppins"}}>
-                <div>
-                    <p className="navBarTitle" style={{display: "inline", color: "white"}}>EnergX</p>
-                    <button className="navbarButton" onClick={logOut}> Log out</button>
-                </div>
-            </Navbar>
-            <ProSidebar style={{
-                marginTop: 48,
-                position: "fixed",
-                fontFamily: "poppins",
-                sticky: "top",
-                width: "150px",
-                minWidth: "150px",
-                fontSize: "20px",
-            }} className="mySideBar">
-                <Menu iconShape="square">
-                    <MenuItem onClick={() => {}}>Menu</MenuItem>
-                    <SubMenu title="Games">
-                        <MenuItem onClick={() => navigate("/admin/users")}> Classic quiz</MenuItem>
-                        <MenuItem onClick={() => {}}> Drag & Drop</MenuItem>
-                    </SubMenu>
-                    <SubMenu title="Courses">
-                        <MenuItem onClick={() => navigate("/admin/devices")}> English</MenuItem>
-                        <MenuItem onClick={() => {}}> IT</MenuItem>
-                    </SubMenu>
-                </Menu>
-            </ProSidebar>
+            <NavbarCustom/>
+            <SidebarCustom/>
             <div className="App" style={{fontFamily:"poppins"}}>
                 <header className="myHeader">
                     <div className="one-below-another">
